dont bother reporting which banks failed
since writes to the empty banks rely on the previous value being retained,
that means that if the previous value is wrong then it will be marked as a
failure.  And since we only reported the write-bank as the failure and not the
read bank, that means that all banks will be listed as failures just because
the previous bank's failure caused the empty bank to fail.